Create DouglasGM_2018_metadata.tsv
Some of the columns need to be updated to fit within the template. For example, there is no country code for Scotland in the template. There is no code for the body site, which is an intestinal biopsy. There is also no code for 'responder vs. nonresponder', which this dataset is concerned with. 
I have to match up the NCBI IDs with the sample IDs. Also, there are no subject IDs in this CSV, so I have to confirm that each sample is 1 subject.