Added an unpublished IBD Dataset from 2012
We found another (to my knowledge unpublished) dataset on SRA containing labelled samples from healthy and IBD patients in Cardiff, UK. I drafted a metadata tsv based on my understanding of the requirements. I haven't done the mimum/median read length yet, but I can do so if the rest of this looks good for inclusion.

The SRA runs for the dataset are located here:

https://www.ncbi.nlm.nih.gov/Traces/study/?query_key=10&WebEnv=MCID_6048f0b09791a01869656a42&o=acc_s%3Aa